Initial commit, add marshalling and something resembling rest
Tavern is a RESTfull toolkit wrapped around cowboy[1]

The main goal is to provide easy handling of different content
types through common a common marshalling structure. There is
aslo a API that can be hooked onto to control what content-types
to handle along with which HTTP verbs to expose and authorization
for those.

[1] -> github.com/extend/cowboy